requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==4.12.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
pip==22.3.1
pycparser==2.22
setuptools==65.5.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
drf_yasg
